more size_t
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5497 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_demo.c
@@ -185,7 +185,7 @@ void CL_ReadDemoMessage(void)
 		VectorCopy(cl.mviewangles[0], cl.mviewangles[1]);
 		for (i = 0;i < 3;i++)
 		{
-			r = FS_Read(cls.demofile, &f, 4);
+			r = (int)FS_Read(cls.demofile, &f, 4);
 			cl.mviewangles[0][i] = LittleFloat(f);
 		}
